requirements.txt
python-telegram-bot==21.0.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2